Changed imserver:
1. Timesout if it hasn't heard from the client

2. Now expects a fits table instead of an image. The table consists
of columns x, y, and fwhm these are pixel position of each source
and the corresponding fwhm. The fits file should include the guess for the
ra and dec of the the image. We should later include plate scale information,
for the solver.

3. The response is no longer a simple text giving some important information
now it is a json header followed by each file the solver generates. Notably:
wcs, rdls, fits, corr and match. The json header contains the file size of
of the various files so they can be seperated properly. The files are sent as
a continuous stream of binary data.

The imserver is meant to be run on nimoy so that it can solve as fast as
possible.

I have also added an imclient. Before, I was just using nc to send images to
the to imserver. That has become untenable because of the more complicated
response from the server mentioned above. The imclient creates the x,y,fwhm
binary fits table and sends it to the server. Then it reads the header and
divies the files from the server accordingly.